requirements.txt
opencv-python
moviepy
Pillow
numpy
pygame
requests 